player stats, items

+++ Assets/Source/Scripts/Player/IDamagable.cs
@@ -1,5 +1,5 @@
 public interface IDamagable
 {
-    void TakeDamage(int damage);
-    void RPC_TakeDamage(int viewID, int damage);
+    void TakeDamage(float damage);
+    void RPC_TakeDamage(int viewID, float damage);
 }
